+++ actix-web-actors/CHANGES.md
@@ -1,6 +1,13 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
+* Update `actix` to `0.12`. [#2277]
+
+[#2277]: https://github.com/actix/actix-web/pull/2277
+
+
+## 4.0.0-beta.5 - 2021-06-17
+* No notable changes.
 
 
 ## 4.0.0-beta.4 - 2021-04-02